added field checking with hapi.js, renamed /user to /users
